fake/requirements.txt
black==24.10.0
click==8.1.8
Faker==33.1.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
